// apps/web/src/services/generated/models/userVo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * API Documentation
 *
 *     当前环境: development
 *     最后更新: 9/15/2026, 5:29:26 PM
 *     [项目设计文档](https://wiki.example.com)
 *
 * OpenAPI spec version: 1.0
 */
import type { AuthSystemRoleVo } from "./authSystemRoleVo";
import type { UserBusinessRoleVo } from "./userBusinessRoleVo";

export interface UserVo {
  /** 用户 ID */
  id: string;
  /** 用户名 */
  username: string;
  /**
   * 邮箱
   * @nullable
   */
  email?: string | null;
  /** 所属系统角色 */
  systemRole: AuthSystemRoleVo;
  /** 连续登录失败次数 */
  failedLoginAttempts: number;
  /**
   * 锁定截止时间
   * @nullable
   */
  lockedUntil?: string | null;
  /** 是否禁止登录 */
  disabled: boolean;
  /** 创建时间 */
  createdTime: string;
  /** 更新时间 */
  updatedTime: string;
  /** 已授予的业务角色 */
  businessRoles: UserBusinessRoleVo[];
}